Implement new posts component.
This commit introduces a new post search secondary panel and implements it in a post component type, which should be extended to support more narrow post searching.

It also refactors some of the LESS files and generates one control and one preview style.

The content-block type should be refactored so that the link panel makes use of the new post search secondary panel.