Fix error where shallow copy left references in component models
The array attributes in component models can not be cloned, leading to issues
with changes to a component instance effecting model defaults. This commit
relies on the Backbone models for default values. The clc_components hash
now only includes meta attributes, which typically include the name and
description, but may also include attributes defined procedurally. Meta
attributes should be JS primitives or they will be cloned by reference. If
they are not JS primitives, they should ideally be attributes that aren't
modified for each instance of a component. See the comments added in this
commit for more information.